combine file ulit
